added new functional with permissions
