+++ homework5/src/main/java/ru/itis/homework5/Fragment/MainFragment.kt
@@ -1,10 +1,6 @@
 package ru.itis.homework5.Fragment
 
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import android.util.Log
 import android.view.View
 import android.widget.Toast
@@ -17,12 +13,11 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.core.content.ContextCompat
 import androidx.lifecycle.lifecycleScope
 import by.kirich1409.viewbindingdelegate.viewBinding
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.async
@@ -37,7 +32,7 @@ import ru.itis.homework5.ui.RadioButtonSingleSelection
 import ru.itis.homework5.ui.SimpleOutlinedTextFieldSample
 import ru.itis.homework5.ui.ThreadPool
 
-class ComposeSampleFragment: BaseFragment(R.layout.fragment_compose_sample) {
+class MainFragment: BaseFragment(R.layout.fragment_compose_sample) {
     private var fragmentJob = SupervisorJob()
     private val scope: CoroutineScope = lifecycleScope //от этого scope надо вызывать все корутины и от него останавливать
     val jobsList = mutableListOf<Job>()
@@ -45,7 +40,7 @@ class ComposeSampleFragment: BaseFragment(R.layout.fragment_compose_sample) {
 
     private var count_of_corutines by mutableStateOf("")  //количество введенных корутин
     private var how_launch by mutableStateOf("Sequentially") // последовательно или параллельно
-    private var selectedLogic by mutableStateOf("Cancel") //логика запуска: отменить при сворачивании или нет
+    private var selectedLogic by mutableStateOf("Cancel") //отменить при сворачивании или нет
     private var selectedDispatcher by mutableStateOf("Default")
 
     private val viewBinding: FragmentComposeSampleBinding by viewBinding(FragmentComposeSampleBinding::bind)
@@ -109,29 +104,44 @@ class ComposeSampleFragment: BaseFragment(R.layout.fragment_compose_sample) {
             isError = true
             return
         }
+        isError = false
+
         val dispatchers = when (selectedDispatcher) {
             "Default" -> Dispatchers.Default
             "IO" -> Dispatchers.IO
             "Main" -> Dispatchers.Main
             else -> Dispatchers.Unconfined
         }
-        if (how_launch == "Parallel") {
-            repeat(count_of_corutines.toInt()) { count ->
-                val job = scope.async(dispatchers, start = CoroutineStart.DEFAULT) {
-                    delay(1000L)
-                    Log.i("Courutin is start", "Корутина $count запущена")
-                }
-                jobsList.add(job)
-            }
-        } else if (how_launch == "Sequentially") {
-            scope.launch(dispatchers){
+        val coroutineScope = if (selectedLogic == "Continue") GlobalScope else scope
+
+        when {
+            how_launch == "Parallel" -> {
                 repeat(count_of_corutines.toInt()) { count ->
-                    val job = scope.async {
-                        delay(1000L)
-                        Log.i("Coroutine started", "Корутина $count запущена")
+                    val job = coroutineScope.launch(dispatchers) {
+                        try {
+                            delay(1000L)
+                            Log.i("Courutin is start", "Корутина $count запущена")
+                        } catch (e: Exception) {
+                            Toast.makeText(context, "Ошибка запуска корутин", Toast.LENGTH_SHORT).show()
+                        }
                     }
                     jobsList.add(job)
-                    job.await()
+                }
+            }
+            else -> {
+                coroutineScope.launch(dispatchers){
+                    try {
+                        repeat(count_of_corutines.toInt()) { count ->
+                            val job = scope.async {
+                                delay(1000L)
+                                Log.i("Coroutine started", "Корутина $count запущена")
+                            }
+                            jobsList.add(job)
+                            job.await()
+                        }
+                    }catch (e: Exception) {
+                        Toast.makeText(context, "Ошибка запуска корутин", Toast.LENGTH_SHORT).show()
+                    }
                 }
             }
         }
@@ -142,6 +152,7 @@ class ComposeSampleFragment: BaseFragment(R.layout.fragment_compose_sample) {
             Toast.makeText(context, "Корутины не были запущены", Toast.LENGTH_SHORT).show()
             return
         }
+        Toast.makeText(context, "Количество корутин, которое было отменено: ${jobsList.size}", Toast.LENGTH_SHORT).show()
         Log.i("Количество корутин, которое было отменено", "${jobsList.size}")
         jobsList.forEach { it.cancel() }
         jobsList.clear()
